backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Vector Store
chromadb==0.4.22
pgvector==0.2.5

# AI/ML
torch==2.1.2
sentence-transformers==2.3.1
transformers==4.37.2
scikit-learn==1.4.0
xgboost==2.0.3
pandas==2.2.0
numpy==1.26.3

# Security
cryptography==41.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
openpyxl==3.1.2
Pillow==10.2.0
python-docx==1.1.0
PyPDF2==3.0.1
httpx==0.26.0

# Monitoring
prometheus-client==0.19.0
structlog==24.1.0

# Caching
redis==5.0.1
fastapi-cache2==0.2.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0
